Updated codes for video in RN theorem
