+++ FSF-2020/linear-algebra/linear-transformations/The-Rank-Nullity-Theorem/file.txt
@@ -1,2 +1,3 @@
-file 'RN_Theorem.mp4'
-file 'RN_Theorem1.mp4'
+file 'RN_Line.mp4'
+file 'RN_Point.mp4'
+file 'RN_SameDim.mp4'
