--- FSF-2020/linear-algebra/linear-transformations/The-Rank-Nullity-Theorem/file2_RN_Theorem1.py
@@ -1,33 +1,0 @@
-from manimlib.imports import *
-class RN_Theorem1(LinearTransformationScene):
-    def construct(self):
-        self.setup()
-        self.wait()
-        square = Square(fill_color=BLACK, fill_opacity=1,color = BLUE)
-        square.move_to(4*LEFT+4*DOWN)
-        square.scale(4)
-        self.play(ShowCreation(square))
-        predim = TextMobject("Another One")
-        predim.move_to(DOWN+4*LEFT)
-        predim.scale(0.75)
-        self.play(Write(predim))
-        self.wait()
-        self.play(FadeOut(predim))
-        afterlt = TextMobject("After Linear transformation")
-        afterlt.move_to(DOWN+4*LEFT)
-        afterlt.scale(0.75)
-        afterlt2 = TextMobject("Dimension of the vector space","changes to 0")
-        afterlt2[0].move_to(2*DOWN+4*LEFT)
-        afterlt2[1].move_to(3*DOWN+4*LEFT)
-        afterlt2.scale(0.75)
-        matrix = [[0,0],[0,0]]
-        self.apply_matrix(matrix)
-        self.play(Write(afterlt))
-        self.play(Write(afterlt2))
-        self.wait()
-        nullity = TextMobject("Hence, nullity = 2")
-        nullity.move_to(DOWN+4*LEFT)
-        #afterlt.scale(0.75)
-        self.play(FadeOut(afterlt),FadeOut(afterlt2),Write(nullity))
-        self.wait(1)
-        self.play(FadeOut(nullity))
